requirements.txt
streamlit
PyMuPDF
google-generativeai
python-dotenv
sentence-transformers
faiss-cpu
numpy
SpeechRecognition
pyaudio
